Create Doctrine Types
Create Doctrine DBAL types for each IP value object class, along with a backwards compatible IpType class which emits a E_USER_DEPRECATED error.
