+++ src/Doctrine/IPv6Type.php
@@ -1,0 +1,27 @@
+<?php
+
+namespace Darsyn\IP\Doctrine;
+
+use Darsyn\IP\IPv6 as IP;
+
+/**
+ * {@inheritDoc}
+ */
+class IPv6Type extends AbstractType
+{
+    /**
+     * {@inheritDoc}
+     */
+    protected function getIpClass()
+    {
+        return IP::class;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    protected function createIpObject($ip)
+    {
+        return new IP($ip);
+    }
+}
