+++ src/Doctrine/MultiType.php
@@ -1,0 +1,27 @@
+<?php
+
+namespace Darsyn\IP\Doctrine;
+
+use Darsyn\IP\Multi as IP;
+
+/**
+ * {@inheritDoc}
+ */
+class MultiType extends AbstractType
+{
+    /**
+     * {@inheritDoc}
+     */
+    protected function getIpClass()
+    {
+        return IP::class;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    protected function createIpObject($ip)
+    {
+        return new IP($ip);
+    }
+}
